requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
python-multipart==0.0.6
astrapy>=1.2,<2.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
langchain-text-splitters==0.2.2
langchain-astradb==0.3.3
ollama==0.1.7